client/requirements.txt
pika==1.2.0
grpcio 
grpcio-tools